+++ cli/src/tasks/migrate.ts
@@ -171,6 +171,16 @@ export async function migrateCommand(
         existsSync(config.ios.platformDirAbs)
       ) {
         // ios template changes
+        // Remove NSLocationAlwaysUsageDescription
+        await runTask(
+          `Migrating Info.plist by removing NSLocationAlwaysUsageDescription key.`,
+          () => {
+            return removeKey(
+              join(config.ios.nativeTargetDirAbs, 'Info.plist'),
+              'NSLocationAlwaysUsageDescription',
+            );
+          },
+        );
       }
 
       if (
@@ -183,6 +193,41 @@ export async function migrateCommand(
             variablesAndClasspaths,
           );
         });
+
+        // Replace deprecated compileSdkVersion
+        await runTask(
+          'Replacing deprecated compileSdkVersion from build.gradle',
+          () => {
+            return (async (): Promise<void> => {
+              const buildGradleFilename = join(
+                config.android.platformDirAbs,
+                'app',
+                'build.gradle',
+              );
+              const buildGradleText = readFile(buildGradleFilename);
+
+              if (!buildGradleText) {
+                logger.error(
+                  `Could not read ${buildGradleFilename}. Check its permissions and if it exists.`,
+                );
+                return;
+              }
+              const compileSdk = `compileSdkVersion rootProject.ext.compileSdkVersion`;
+              if (buildGradleText.includes(compileSdk)) {
+                const buildGradleReplaced = buildGradleText.replace(
+                  compileSdk,
+                  `compileSdk rootProject.ext.compileSdkVersion`,
+                );
+
+                writeFileSync(
+                  buildGradleFilename,
+                  buildGradleReplaced,
+                  'utf-8',
+                );
+              }
+            })();
+          },
+        );
 
         // Update gradle-wrapper.properties
         await runTask(
@@ -248,10 +293,10 @@ export async function migrateCommand(
               }
             }
             const pluginVariables: { [key: string]: string } = {
-              firebaseMessagingVersion: '23.2.1',
+              firebaseMessagingVersion: '23.3.1',
               playServicesLocationVersion: '21.0.1',
-              androidxBrowserVersion: '1.5.0',
-              androidxMaterialVersion: '1.9.0',
+              androidxBrowserVersion: '1.7.0',
+              androidxMaterialVersion: '1.10.0',
               androidxExifInterfaceVersion: '1.3.6',
               androidxCoreKTXVersion: '1.12.0',
               googleMapsPlayServicesVersion: '18.1.0',
@@ -405,9 +450,10 @@ async function installLatestLibs(
 async function writeBreakingChanges() {
   const breaking = [
     '@capacitor/camera',
-    '@capacitor/device',
+    '@capacitor/filesystem',
+    '@capacitor/geolocation',
+    '@capacitor/google-maps',
     '@capacitor/local-notifications',
-    '@capacitor/push-notifications',
   ];
   const broken = [];
   for (const lib of breaking) {
@@ -417,7 +463,7 @@ async function writeBreakingChanges() {
   }
   if (broken.length > 0) {
     logger.info(
-      `IMPORTANT: Review https://capacitorjs.com/docs/updating/5-0#plugins for breaking changes in these plugins that you use: ${broken.join(
+      `IMPORTANT: Review https://capacitorjs.com/docs/next/updating/6-0#plugins for breaking changes in these plugins that you use: ${broken.join(
         ', ',
       )}.`,
     );
@@ -606,9 +652,6 @@ async function updateBuildGradle(
     'com.google.gms:google-services': string;
   },
 ) {
-  // In build.gradle add dependencies:
-  // classpath 'com.android.tools.build:gradle:8.0.0'
-  // classpath 'com.google.gms:google-services:4.3.15'
   const txt = readFile(filename);
   if (!txt) {
     return;
@@ -758,4 +801,29 @@ export async function patchOldCapacitorPlugins(
       }
     }),
   );
+}
+
+async function removeKey(filename: string, key: string) {
+  const txt = readFile(filename);
+  if (!txt) {
+    return;
+  }
+  let lines = txt.split('\n');
+  let removed = false;
+  let removing = false;
+  lines = lines.filter(line => {
+    if (removing && line.includes('</string>')) {
+      removing = false;
+      return false;
+    }
+    if (line.includes(`<key>${key}</key`)) {
+      removing = true;
+      removed = true;
+    }
+    return !removing;
+  });
+
+  if (removed) {
+    writeFileSync(filename, lines.join('\n'), 'utf-8');
+  }
 }